i solve some bugs in exam part and in admin amd principal login part and combobox label. but login in exam part have bug yet
