+++ app/layout.tsx
@@ -1,35 +1,27 @@
+"use client";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import { useAuth } from "@/hooks/useAuth";
+import { useEffect } from "react";
 
-<link href="https://fonts.googleapis.com/css2?family=Vazirmatn:wght@400;500;700&display=swap" rel="stylesheet" />
-
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const { setupAxiosInterceptor, apiClient } = useAuth();
+
+  useEffect(() => {
+    // تنظیم interceptor هنگام بارگذاری برنامه
+    setupAxiosInterceptor(apiClient);
+  }, []);
   return (
     <html lang="fa" dir="rtl">
 
       <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+        
       >
         {children}
       </body>
